[seller] refactor: 중복 정의된 checkStoreName을 entity/store로 통합
