+++ apps/seller/src/entity/seller-info/seller-info.type.ts
@@ -1,19 +1,4 @@
-export interface Store {
-  storeId: number
-  name: string
-  introduce: string
-  profile: string
-  phoneNumber: string
-  subPhoneNumber: string
-  email: string
-  originAddress: string
-  originAddressDetail: string
-}
-
-export interface StoreNameCheckResult {
-  available: boolean
-  store: Store | null
-}
+import type { Store } from '@/entity/store'
 
 export interface MyStoreSummary {
   // 변경 신청 가능 여부: 이미 변경했거나 PENDING 이면 false
