+++ apps/seller/src/entity/store/index.ts
@@ -1,0 +1,2 @@
+export { checkStoreName } from './store.api'
+export type { Store, StoreNameCheckResult } from './store.type'
