+++ apps/seller/src/entity/store/store.type.ts
@@ -1,0 +1,16 @@
+export interface Store {
+  storeId: number
+  name: string
+  introduce: string
+  profile: string
+  phoneNumber: string
+  subPhoneNumber: string
+  email: string
+  originAddress: string
+  originAddressDetail: string
+}
+
+export interface StoreNameCheckResult {
+  available: boolean
+  store: Store | null
+}
