+++ apps/seller/src/features/register/store-info/check-store-name.mutation.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query'
 
-import { checkStoreName } from '@/entity/register'
+import { checkStoreName } from '@/entity/store'
 
 export const useCheckStoreNameMutation = () => {
   return useMutation({
